# fix(cli): fix cache not being used when there are no dependencies