--- a/change/@react-native-windows-cli-401a1aee-fc62-4128-b881-42b970e6102e.json
+++ b/change/@react-native-windows-cli-401a1aee-fc62-4128-b881-42b970e6102e.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "Fix cache not being used when there are no dependencies",
+  "packageName": "@react-native-windows/cli",
+  "email": "4123478+tido64@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}