## Description

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
This addresses a perf regression in autolinking when there are no dependencies to link.

Resolves #9518

### What
Instead of checking whether the dependencies cache is empty, check whether it was instantiated.

## Testing
| Before | After |
| -: | -: |
| 22488ms | 3561ms |