@tido64 2022-02-11
Looks like Beachball is having issues again:

```
ERROR: Attempting to bump to a version that already exists in the registry: @office-iss/react-native-win32@0.0.0-canary.133

ERROR: Attempting to bump to a version that already exists in the registry: @react-native-windows/virtualized-list@0.0.0-canary.29
Validating package version - react-native-windows-init@1.1.86 OK!
Validating package version - react-native-windows@0.0.0-canary.454 OK!
Validating package version - @react-native-windows/automation-channel@0.1.28 OK!
Validating package version - @react-native-windows/automation@0.1.39 OK!
Validating package version - @react-native-windows/automation-commands@0.0.66 OK!
Something went wrong with the publish! Manually update these package and versions:
- @react-native-windows/cli@0.0.0-canary.115
- react-native-windows-init@1.1.86
- react-native-windows@0.0.0-canary.454
- @react-native-windows/automation-channel@0.1.28
- @react-native-windows/automation@0.1.39
- @react-native-windows/automation-commands@0.0.66
- e2e-test-app@0.0.0
- integration-test-app@0.0.0
- playground@0.0.54
- sample-apps@0.0.0
- @office-iss/react-native-win32@0.0.0-canary.133
- @office-iss/react-native-win32-tester@0.0.1
- @react-native-windows/tester@0.0.1
- @react-native-windows/virtualized-list@0.0.0-canary.29
```