@NickGerleman 2022-02-11
Seeing the same Beachball issue with https://github.com/microsoft/react-native-windows/pull/9491. Doing a `beachball bump --verbose` on a local branch merged to main, it seems to be calculating some packages to be bumped that are already up-to-date (with package.json up to date as well). `beachball sync` sees that everything locally is correct as well. So something seems to be going wrong during the actual bumping process evaluation I think?